created server-side caching for getting owned playlists (GET playlists) functionality

+++ src/services/postgres/playlistsService.js
@@ -5,9 +5,9 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistsService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
-
+    this._cacheService = cacheService;
     this._collaborationService = collaborationService;
   }
 
@@ -25,22 +25,28 @@ class PlaylistsService {
       throw new InvariantError('Playlist gagal ditambahkan');
     }
 
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(userId) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username
-      FROM playlists
-      FULL JOIN users ON playlists.owner = users.id
-      FULL JOIN collaborations ON playlists.id = collaborations.playlist_id
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-      values: [userId],
-    };
+    try {
+      const result = await this._cacheService.get(`playlists:${userId}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username
+        FROM playlists
+        FULL JOIN users ON playlists.owner = users.id
+        FULL JOIN collaborations ON playlists.id = collaborations.playlist_id
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+        values: [userId],
+      };
 
-    const result = await this._pool.query(query);
-
-    return result.rows;
+      const result = await this._pool.query(query);
+      await this._cacheService.set(`playlists:${userId}`, JSON.stringify(result.rows));
+      return result.rows;
+    }
   }
 
   async verifyPlaylistOwner(id, owner) {
@@ -78,14 +84,17 @@ class PlaylistsService {
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
     const result = await this._pool.query(query);
+
     if (!result.rows.length) {
       throw new NotFoundError('Maaf, playlist yang anda ingin hapus tidak ditemukan');
     }
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 }
